split http_client
