+++ evalscope/perf/plugin/__init__.py
@@ -1,0 +1,2 @@
+from evalscope.perf.plugin.api import *
+from evalscope.perf.plugin.datasets import *
