+++ evalscope/perf/plugin/api/__init__.py
@@ -1,0 +1,3 @@
+from evalscope.perf.plugin.api.custom_api import CustomPlugin
+from evalscope.perf.plugin.api.dashscope_api import DashScopeApiPlugin
+from evalscope.perf.plugin.api.openai_api import OpenaiPlugin
