+++ evalscope/perf/plugin/datasets/__init__.py
@@ -1,0 +1,3 @@
+from evalscope.perf.plugin.datasets.line_by_line import LineByLineDatasetPlugin
+from evalscope.perf.plugin.datasets.longalpaca import LongAlpacaDatasetPlugin
+from evalscope.perf.plugin.datasets.openqa import OpenqaDatasetPlugin
